fix: move category products to a separate component
